Add files via upload
This Dashboard is not pandemic specific, but rather a remote workforce security analysis based on impossible VPNs, VPN outside a specific country, find VPNs for a user, possible phishing using malformed links impersonating Teams, Zoom, etc. It also grants the ability to specify a time, reducing the number of logs Splunk is set to parse or even the ability to examine "after business hours" groups.